Update api client
I fixed some super weird behaviour when languages changed in between requests. However, it changes the internal behaviour for extending endpoints and is therefore a breaking release. I also removed one of the cache storages (`localForage`) that proved to be extremely slow.

Nothing of this matters here since you don't use any of it, but you still need to bump the version. :smile: