make sure css is written before continuing with next buildstep
modular-css/browserify is waiting for the 'bundle' event of browserify, and then async writing files after the stream 'end' event. Because the end event has no callback or similar, browserify continues with the next buildstep while modular-css is still processing all the files (`processor.output()`). The 'revision' task is running before the css is written, resulting in the css file not being copied to the public directory.

This delays the next step in the gulp chain until the modular-css 'done' plugins are run.